Änderung Rechtesystem, refactoring und bugfixes. Unbedingt Datenbank updaten!
Zwei neue Rechte sind hinzugefügt worden, rechtesystem in eigene Datei ausgelagert umd auch im Frontend darauf zugreifen zu können. Einige codefixes.
