+++ admin/dashboard/settings/settings_form.php
@@ -1,0 +1,10 @@
+<div class="row d-flex flex-column">
+    <div class="col my-3">
+        <h4 class="text-center">ClanMS - Systemeinstellungen</h4>
+        <hr />
+    </div>
+</div>
+<?php 
+    include_once(__DIR__."/settings_homepage.php");
+    include_once(__DIR__."/settings_owner.php");
+?>
